Flex tab in channel editor
